Command-level rights checked on GLPI side
Commands you don't have access to are hidden and refused to be sent through Jamf API.
Had to add the ability to select return format for a GET API call because only one group for a JSS account was returned.
Had to add extra API calls when getting rights for a user with Group-level access because "jss_actions" are not being returned with the user.